[FIX] website: don't interpolate null values for css properties
In 6d8de09073c8 we bumped the verion of libsass that we use. One
breaking change introduced by the new version is that if a css variable
is interpolated to the empty string, it is still emited in the compiled
css [1]

This commit fixes that by explicitly omitting null values from the map
before looping over it.

[1]: https://sass-lang.com/documentation/breaking-changes/css-vars

task-2248360

closes odoo/odoo#50668

X-original-commit: 6faf327706169bbe31e188a42805085b95b78052
Signed-off-by: Quentin Smetz (qsm) <qsm@odoo.com>