[FIX] survey: redirect on login page on survey start if not logged in.
Before this commit, when sharing a survey 'Login required' to a non logged in
user, the redirection was made to the home page.
The reason is: when the survey is not started yet, no user input is already
created for this survey and this user. So if no user_input has been created,
yet, that means that the user has never started the survey before and user
need first to login before continuing.

After this commit, we redirect to the loggin page.

Task ID: 2246430

closes odoo/odoo#58965

X-original-commit: e563850fe18cee15b4e4e7cd5380c3664ba7cf59
Signed-off-by: Thibault Delavallee (tde) <tde@openerp.com>
Signed-off-by: David Beguin <dbeguin@users.noreply.github.com>