[FIX] web: force refresh on expired session popup button click
Currently, when the session expired then it generates the popup that
"Your odoo session expired. Please refresh the current page."
Now once you click on the "Ok" button only close the popup and user have to
manually do refresh the page.

with this commit, we have changed the text of the popup to
"Your odoo session expired. The current page is about to be refreshed." and
when user click on "Ok" button forcefully refresh the page.

Task : 2241744

closes odoo/odoo#50710

Signed-off-by: Aaron Bohy (aab) <aab@odoo.com>