[FIX] point_of_sale: wrong date in pos payment
- Have a user in an early timezone (for example Samoa Standard Time);
- Open a POS session in the evening (say 20h).
- Buy something and pay with cash.
- Close the session.
- List the session orders, and open the newly create order.

Before this commit, the date of the order was correct (today at 20h),
but the date of the payment was wrong (yesterday at 13h). This error
occurs because a wrong format was used and the date was taken into
account two times the timezone.

Now, both dates are corrected.

This fix is a partial revert of d666e5f6ce01ef152615ca39f19a8a2fca43b8b8

opw-2231820

X-original-commit: 04853fcd33a6d25727fcfd4d982f96106804e94b