[FIX] tools: reimport translations in csv
Before this commit trying to reimport csv of a translation failed.
The reason was that before 53fee36f70eccda the parsing was common
between po and csv while now it's split in two different parsers.
The res_id column of the CSV is the external id of the record and is
handled by IrTranslationImport.

Use DictReader instead of csv_reader to easily add new keys and still
be flexible on the given csv files.

Match the POReader format with a imd_model and imd_name column.

As for the PoFileReader, the code translations are unique and must be
discarded in case of duplicate.

Correct the error message if the imported file is not correct (was
missing an argument)

Fixes odoo/odoo#50975

closes odoo/odoo#51468

X-original-commit: 996a3b0058000f21b57332e2b4052415bd4d0d2a
Signed-off-by: Martin Trigaux (mat) <mat@odoo.com>