[FIX] web_editor: ease input selection in the left panel
If a mousedown event was fired over the left panel in edit mode while
the mouse up was fired over the current page in edit, the focus would
suddenly go at the top of the page due to pointer events. This made
input selection not easy on some browser.

task-2210712

closes odoo/odoo#52495

X-original-commit: 8110f60cbb23e40a25510559ec9e383097d93303
Signed-off-by: Quentin Smetz (qsm) <qsm@odoo.com>