[IMP] stock: show location name in warning
Before this commit if you have a quant that has a location linked of the type 'view' you'd get this warning
but you would not know about which location this was and figuring that out is not trivial for most people.
After this commit users get notified about which location is of the type 'view' and can directly correct it if they want to.

closes odoo/odoo#51137

Signed-off-by: Simon Lejeune (sle) <sle@openerp.com>